Refactor access control and error handling
Enhanced access control and error handling across multiple services and controllers. Introduced `CustomResult` for better error management. Key changes include:

- Added `SystemAccountEmail` constant in `Constants.cs`.
- Introduced `ToNewTimeLineItem` extension method in `CalendarItemExtensions.cs`.
- Removed `CopyCalendarItemPropertiesForAdd` in `TimeLineItemExtensions.cs`.
- Changed error code constants to `public static readonly` in `CustomError.cs`.
- Updated `GetCalendarList` and `GetProgenyUserAccessList` method calls to include access level and user email parameters.
- Refactored `UserInfoController.cs` to improve security by using `POST` for email-based user info retrieval.
- Updated `AzureNotifications.cs` and various services to return `CustomResult` for better error handling.
- Added `ResultExtensions.cs` for utility methods to handle `CustomResult` objects.
- Enhanced `FamilyController.cs` to include `IUserInfosHttpClient` and populate user access objects with necessary details.

+++ KinaUnaProgenyApi/Controllers/ProgenyController.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using KinaUna.Data;
 using KinaUna.Data.Extensions;
 using KinaUna.Data.Models;
+using KinaUna.Data.Models.DTOs;
 using KinaUnaProgenyApi.Services;
 using KinaUnaProgenyApi.Services.UserAccessService;
 using Microsoft.AspNetCore.Authorization;
@@ -255,17 +256,17 @@ namespace KinaUnaProgenyApi.Controllers
                 return Unauthorized();
             }
             
-            List<UserAccess> userAccessList = await userAccessService.GetProgenyUserAccessList(progeny.Id);
+            CustomResult<List<UserAccess>> userAccessListResult = await userAccessService.GetProgenyUserAccessList(progeny.Id, userEmail);
+            if (!userAccessListResult.IsSuccess) return userAccessListResult.ToActionResult();
 
-            if (userAccessList.Count != 0)
+            foreach (UserAccess ua in userAccessListResult.Value)
             {
-                foreach (UserAccess ua in userAccessList)
-                {
-                    await userAccessService.RemoveUserAccess(ua.AccessId, ua.ProgenyId, ua.UserId);
-                }
+                await userAccessService.RemoveUserAccess(ua.AccessId, ua.ProgenyId, ua.UserId);
             }
 
             await progenyService.DeleteProgeny(progeny);
+
+
             return NoContent();
 
         }
